+++ pkg/repositories/repositories.go
@@ -35,10 +35,18 @@ type ListResult struct {
 	Repositories  []Repo
 }
 
+// GithubRepo is a sparce struct for just the GitHub repository info we need.
+type GithubRepo struct {
+	Name        string
+	Description string
+	IsPrivate   bool
+}
+
 type Service struct {
 	config       Config
 	ctx          context.Context
 	githubClient *github.Client
+	githubRepos  map[string]GithubRepo
 }
 
 // New instantiates a new repositories service.
@@ -67,22 +75,61 @@ func New(c Config) (*Service, error) {
 		githubClient: client,
 	}
 
+	repos, err := s.loadGithubRepoData()
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	s.githubRepos = repos
+
 	return s, nil
 }
 
-// LoadList loads a list of repository configurations from a local path.
+// Load repository metadata from Github.
+func (s *Service) loadGithubRepoData() (map[string]GithubRepo, error) {
+	opts := &github.RepositoryListByOrgOptions{
+		ListOptions: github.ListOptions{
+			PerPage: 100,
+		},
+	}
+	repos := make(map[string]GithubRepo)
+
+	for {
+		r, resp, err := s.githubClient.Repositories.ListByOrg(s.ctx, s.config.GithubOrganization, opts)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, repo := range r {
+			repos[repo.GetName()] = GithubRepo{
+				Name:        repo.GetName(),
+				Description: repo.GetDescription(),
+				IsPrivate:   repo.GetPrivate(),
+			}
+		}
+
+		if resp.NextPage == 0 {
+			break
+		}
+		opts.Page = resp.NextPage
+	}
+
+	return repos, nil
+}
+
+// Loads a list of repository configurations from a local path.
 // The file name is asserted in the format `<team_name>.yaml`, with all
 // repositories mentioned in it belonging to the team of that name.
-func (s *Service) LoadList(path string) ([]Repo, error) {
+func (s *Service) loadList(path string) ([]Repo, error) {
 	data, err := os.ReadFile(path)
 	if err != nil {
 		return nil, err
 	}
 
-	return s.LoadListFromBytes(data)
+	return s.loadListFromBytes(data)
 }
 
-func (s *Service) LoadListFromBytes(data []byte) ([]Repo, error) {
+func (s *Service) loadListFromBytes(data []byte) ([]Repo, error) {
 	repos := []Repo{}
 	err := yaml.UnmarshalStrict(data, &repos)
 	if err != nil {
@@ -94,6 +141,7 @@ func (s *Service) LoadListFromBytes(data []byte) ([]Repo, error) {
 
 // GetLists loads the lists of repository YAML files from GitHub giantswarm/github.
 func (s *Service) GetLists() ([]ListResult, error) {
+	// Get repositories directory content.
 	_, directoryContent, _, err := s.githubClient.Repositories.GetContents(s.ctx, s.config.GithubOrganization, s.config.GithubRepositoryName, s.config.DirectoryPath, nil)
 	if err != nil {
 		return nil, err
@@ -106,13 +154,14 @@ func (s *Service) GetLists() ([]ListResult, error) {
 			continue
 		}
 
+		// Get individual team repositories file.
 		fileContent, _, _, err := s.githubClient.Repositories.GetContents(s.ctx, s.config.GithubOrganization, s.config.GithubRepositoryName, *item.Path, nil)
 		if err != nil {
 			return nil, err
 		}
 
 		decodedContent, _ := b64.StdEncoding.DecodeString(*fileContent.Content)
-		lists, err := s.LoadListFromBytes(decodedContent)
+		lists, err := s.loadListFromBytes(decodedContent)
 		if err != nil {
 			return nil, err
 		}
@@ -124,4 +173,22 @@ func (s *Service) GetLists() ([]ListResult, error) {
 	}
 
 	return result, nil
+}
+
+// Returns the description for the given repo. If not available,
+// returns an empty string.
+func (s *Service) GetDescription(name string) string {
+	if repo, ok := s.githubRepos[name]; ok {
+		return repo.Description
+	}
+	return ""
+}
+
+// Returns the public/private info for the given repo. If not available,
+// return an error.
+func (s *Service) GetIsPrivate(name string) (bool, error) {
+	if repo, ok := s.githubRepos[name]; ok {
+		return repo.IsPrivate, nil
+	}
+	return false, microerror.Maskf(repositoryNotFoundError, "repository %s not found", name)
 }