+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Export repository descriptions to component entities.
+- Export repository visibility (private) to component entity tags.
+
 ## [0.0.5] - 2023-07-26
 
 ### Changed